fire_dept_automation/ocr_system/requirements.txt
# OCR 文件比對系統 依賴套件

# 核心套件
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# PDF 處理
PyMuPDF>=1.23.0

# OCR 引擎
paddlepaddle-gpu>=2.5.0  # GPU 版本，CPU 版用 paddlepaddle
paddleocr>=2.7.0

# Word 文件處理
python-docx>=1.0.0

# Excel 處理
pandas>=2.0.0
openpyxl>=3.1.0

# HTTP 請求 (LLM API)
requests>=2.31.0

# 選用：OpenAI 相容 API
# openai>=1.0.0
